requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysqlclient
python-dotenv
Pillow